Cleanup solution files
