+++ src/ZeroLog.sln
@@ -9,9 +9,11 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ZeroLog.Tests", "ZeroLog.Te
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Misc", "Misc", "{44EDCA3A-DBC5-415C-9A34-C742912674B5}"
 	ProjectSection(SolutionItems) = preProject
+		..\build\build.cake = ..\build\build.cake
 		..\LICENSE.md = ..\LICENSE.md
 		..\README.md = ..\README.md
 		..\version.yml = ..\version.yml
+		..\build\ZeroLog.nuspec = ..\build\ZeroLog.nuspec
 	EndProjectSection
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ZeroLog.Benchmarks", "ZeroLog.Benchmarks\ZeroLog.Benchmarks.csproj", "{D8CB63CF-F35E-4D38-88AF-9FB4C6B29630}"
