+++ src/ZeroLog/Log.tt
@@ -3,7 +3,6 @@
 <#@ import namespace="System.Linq" #>
 <#@ import namespace="System.Text" #>
 <#@ import namespace="System.Collections.Generic" #>
-<#@ assembly name="$(TargetPath)" #>
 <#@ output extension=".cs" #>
 
 using System;
